左下角标签信息显示框的面积
标定数据精度提升

+++ anylabeling/views/labeling/widgets/canvas.py
@@ -62,7 +62,7 @@ class Canvas(
     # when the cursor moves over an object
     mode_changed = QtCore.pyqtSignal()
     new_shape = QtCore.pyqtSignal()
-    show_shape = QtCore.pyqtSignal(int, int, QtCore.QPointF)
+    show_shape = QtCore.pyqtSignal(float, float, QtCore.QPointF)
     selection_changed = QtCore.pyqtSignal(list)
     shape_moved = QtCore.pyqtSignal()
     shape_rotated = QtCore.pyqtSignal()
@@ -2032,8 +2032,8 @@ class Canvas(
                 return
 
             if self.create_mode in ["rectangle", "cuboid"]:
-                shape_width = int(abs(self.current[0].x() - pos.x()))
-                shape_height = int(abs(self.current[0].y() - pos.y()))
+                shape_width = abs(self.current[0].x() - pos.x())
+                shape_height = abs(self.current[0].y() - pos.y())
                 self.show_shape.emit(shape_height, shape_width, pos)
 
             color = QtGui.QColor(0, 0, 255)
@@ -2148,8 +2148,8 @@ class Canvas(
                 if self.h_shape.shape_type == "rectangle":
                     p1 = self.h_shape[0]
                     p2 = self.h_shape[2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
                 elif (
                     self.h_shape.shape_type == "cuboid"
@@ -2157,8 +2157,8 @@ class Canvas(
                 ):
                     p1 = self.h_shape[0]
                     p2 = self.h_shape[2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
             elif (
                 self.selected_cuboid_face()
@@ -2176,8 +2176,8 @@ class Canvas(
                 self.moving_shape = True
                 p1 = self.h_shape[0]
                 p2 = self.h_shape[2]
-                shape_width = int(abs(p2.x() - p1.x()))
-                shape_height = int(abs(p2.y() - p1.y()))
+                shape_width = abs(p2.x() - p1.x())
+                shape_height = abs(p2.y() - p1.y())
                 self.show_shape.emit(shape_height, shape_width, pos)
             elif self.selected_shapes and self.prev_point:
                 self.h_cuboid_face = None
@@ -2193,8 +2193,8 @@ class Canvas(
                 if self.selected_shapes[-1].shape_type == "rectangle":
                     p1 = self.selected_shapes[-1][0]
                     p2 = self.selected_shapes[-1][2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
                 elif (
                     self.selected_shapes[-1].shape_type == "cuboid"
@@ -2202,8 +2202,8 @@ class Canvas(
                 ):
                     p1 = self.selected_shapes[-1][0]
                     p2 = self.selected_shapes[-1][2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
             else:
                 if (
@@ -2239,8 +2239,8 @@ class Canvas(
                 if self.h_shape.shape_type == "rectangle":
                     p1 = self.h_shape[0]
                     p2 = self.h_shape[2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
                 elif (
                     self.h_shape.shape_type == "cuboid"
@@ -2248,8 +2248,8 @@ class Canvas(
                 ):
                     p1 = self.h_shape[0]
                     p2 = self.h_shape[2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
             elif (
                 self.selected_cuboid_face()
@@ -2266,8 +2266,8 @@ class Canvas(
                 self.moving_shape = True
                 p1 = self.h_shape[0]
                 p2 = self.h_shape[2]
-                shape_width = int(abs(p2.x() - p1.x()))
-                shape_height = int(abs(p2.y() - p1.y()))
+                shape_width = abs(p2.x() - p1.x())
+                shape_height = abs(p2.y() - p1.y())
                 self.show_shape.emit(shape_height, shape_width, pos)
             else:
                 self.is_move_editing = False
@@ -2481,14 +2481,14 @@ class Canvas(
                 if shape.shape_type == "rectangle":
                     p1 = self.h_shape[0]
                     p2 = self.h_shape[2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
                 elif shape.shape_type == "cuboid" and len(self.h_shape) >= 4:
                     p1 = self.h_shape[0]
                     p2 = self.h_shape[2]
-                    shape_width = int(abs(p2.x() - p1.x()))
-                    shape_height = int(abs(p2.y() - p1.y()))
+                    shape_width = abs(p2.x() - p1.x())
+                    shape_height = abs(p2.y() - p1.y())
                     self.show_shape.emit(shape_height, shape_width, pos)
                 break
         else:  # Nothing found, clear highlights, reset state.
